WIP: Removed RunnerOptions and implemented in Runner class itself
